+++ packages/entities/entities-certificates/src/components/CertificateList.vue
@@ -15,6 +15,7 @@
       @clear-search-input="clearFilter"
       @click:row="(row: any) => rowClick(row as EntityRow)"
       @sort="resetPagination"
+      @state="handleStateChange"
     >
       <!-- Filter -->
       <template #toolbar-filter>
@@ -29,20 +30,58 @@
           :disabled="!useActionOutside"
           to="#kong-ui-app-page-header-action-button"
         >
-          <PermissionsWrapper :auth-function="() => canCreate()">
-            <!-- Hide Create button if table is empty -->
+          <div class="button-row">
             <KButton
-              appearance="primary"
-              data-testid="toolbar-add-certificate"
-              :size="useActionOutside ? 'medium' : 'large'"
-              :to="config.createRoute"
+              v-if="enableV2EmptyStates && !showEmptyState && config.app === 'konnect'"
+              appearance="secondary"
+              class="open-learning-hub"
+              data-testid="certificates-learn-more-button"
+              icon
+              @click="$emit('click:learn-more')"
             >
-              <AddIcon />
-              {{ t('certificates.list.toolbar_actions.new_certificate') }}
+              <BookIcon decorative />
             </KButton>
-          </PermissionsWrapper>
+            <PermissionsWrapper :auth-function="() => canCreate()">
+              <!-- Hide Create button if table is empty -->
+              <KButton
+                appearance="primary"
+                data-testid="toolbar-add-certificate"
+                :size="useActionOutside ? 'medium' : 'large'"
+                :to="config.createRoute"
+              >
+                <AddIcon />
+                {{ t('certificates.list.toolbar_actions.new_certificate') }}
+              </KButton>
+            </PermissionsWrapper>
+          </div>
         </Teleport>
       </template>
+
+      <template
+        v-if="enableV2EmptyStates && config.app === 'konnect'"
+        #empty-state
+      >
+        <EntityEmptyState
+          :action-button-text="t('certificates.list.empty_state_v2.create')"
+          appearance="secondary"
+          :can-create="() => canCreate()"
+          :description="t('certificates.list.empty_state_v2.description')"
+          :learn-more="config.app === 'konnect'"
+          :title="t('certificates.list.empty_state_v2.title')"
+          @click:create="handleCreate"
+          @click:learn-more="$emit('click:learn-more')"
+        >
+          <template #image>
+            <div class="empty-state-icon-gateway">
+              <ServiceDocumentIcon
+                :color="KUI_COLOR_TEXT_DECORATIVE_AQUA"
+                :size="KUI_ICON_SIZE_50"
+              />
+            </div>
+          </template>
+        </EntityEmptyState>
+      </template>
+
 
       <!-- Column Formatting -->
       <template #subject="{ row }">
@@ -156,11 +195,14 @@
 <script setup lang="ts">
 import type { PropType } from 'vue'
 import { computed, ref, watch, onBeforeMount } from 'vue'
+import { KUI_ICON_SIZE_50, KUI_COLOR_TEXT_DECORATIVE_AQUA } from '@kong/design-tokens'
+import { EntityEmptyState } from '@kong-ui-public/entities-shared'
 import type { AxiosError } from 'axios'
 import { useRouter } from 'vue-router'
-import { AddIcon } from '@kong/icons'
+import { AddIcon, BookIcon, ServiceDocumentIcon } from '@kong/icons'
 import composables from '../composables'
 import endpoints from '../certificates-endpoints'
+import '@kong-ui-public/entities-shared/dist/style.css'
 
 import {
   EntityBaseTable,
@@ -172,6 +214,7 @@ import {
   useAxios,
   useFetcher,
   useDeleteUrlBuilder,
+  useTableState,
   TableTags,
 } from '@kong-ui-public/entities-shared'
 import type {
@@ -191,6 +234,7 @@ import '@kong-ui-public/entities-shared/dist/style.css'
 
 const emit = defineEmits<{
   (e: 'error', error: AxiosError): void,
+  (e: 'click:learn-more'): void,
   (e: 'copy:success', payload: CopyEventPayload): void,
   (e: 'copy:error', payload: CopyEventPayload): void,
   (e: 'delete:success', certificate: EntityRow): void,
@@ -249,12 +293,22 @@ const props = defineProps({
     type: Boolean,
     default: false,
   },
+  /**
+   * Enables the new empty state design, this prop can be removed when
+   * the khcp-14756-empty-states-m2 FF is removed.
+   */
+  enableV2EmptyStates: {
+    type: Boolean,
+    default: false,
+  },
 })
 
 const { i18n: { t, formatUnixTimeStamp }, i18nT } = composables.useI18n()
 const router = useRouter()
 
 const { axiosInstance } = useAxios(props.config?.axiosRequestConfig)
+const { hideTableToolbar: showEmptyState, handleStateChange } = useTableState(() => filterQuery.value)
+
 
 /**
  * Table Headers
@@ -469,6 +523,13 @@ const confirmDelete = async (): Promise<void> => {
 }
 
 /**
+ * Create Certificate
+ */
+const handleCreate = (): void => {
+  router.push(props.config.createRoute)
+}
+
+/**
  * Watchers
  */
 watch(fetcherState, (state) => {
@@ -512,6 +573,12 @@ onBeforeMount(async () => {
 </script>
 
 <style lang="scss" scoped>
+.button-row {
+  align-items: center;
+  display: flex;
+  gap: $kui-space-50;
+}
+
 .kong-ui-entities-certificates-list {
   width: 100%;