+++ packages/analytics/analytics-geo-map/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.73](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@0.2.72...@kong-ui-public/analytics-geo-map@0.2.73) (2025-02-11)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
 ## [0.2.72](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@0.2.71...@kong-ui-public/analytics-geo-map@0.2.72) (2025-01-27)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-geo-map