+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.6.20](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@8.6.19...@kong-ui-public/analytics-metric-provider@8.6.20) (2025-02-11)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
 ## [8.6.19](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@8.6.18...@kong-ui-public/analytics-metric-provider@8.6.19) (2025-01-27)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-metric-provider